Tune the playbook for manually setting up a brand new controller
The following allows a stock CentOS 7 server to be launched and
configured as a Slurm controller. As part of the playbook run, the
playbook itself will get copied over and can then be run from the
controller to add workers into cluster (it will configure the worker
and reconfigure the master - the only change required is in the
inventory).