GTN:GA4GH TRS endpoint support rewrite
GA4GH TRS endpoints expect paths ending in nothing, no trailing slash. They additionally have the structure

/blah/1
/blah/1/GALAXY/descriptor

Which obviously doesn't work since we aren't responding to an exact URL but trying to map it to a file. So internally we will call the files 1.json and descriptor.json, and then have nginx strip that out so it just all works