Create a redirect loop to usegalaxy.eu
During the 11th and 12th of June 2023, visitors of usegalaxy.eu were redirected to a blog post (because of maintenance operations), but using a 301 HTTP redirect. This kind of redirect is cached forever by most browsers. Those users need now to clear the browser's cache to access usegalaxy.eu again, but if the news do not reach them, or they do not know how to clear the cache, then they will think that usegalaxy.eu is down forever. As a a workaround, this directive creates a redirection loop, which should invalidate the cache for most browsers.